requirements.txt
lxml==6.0.2
PyMuPDF==1.26.6
python-docx==1.2.0
typing_extensions==4.15.0
